feat(redesignmerge): Move createKeySignatures to new (routes) folder. Remove unneeded code from theme.js, but change typography to Poppins as global font.
