+++ app/theme.js
@@ -2,6 +2,7 @@
 import { createTheme } from "@mui/material";
 import Poppins from "@fontsource/poppins";
 import tokens from "./tokens.json";
+import Montserrat from "typeface-montserrat";
 
 const theme = createTheme({
   palette: {
@@ -29,108 +30,10 @@ const theme = createTheme({
       unfilled: tokens.global.md.sys.color["unfilled-progress-bar-color"].value,
     },
   },
-  // typography: {
-  //   fontFamily: "Poppins",
-  //   sectionTitle: {
-  //     fontFamily:
-  //       tokens.global.md.sys.typescale["section-title"].value.fontFamily,
-  //     fontWeight: 500,
-  //     fontSize: `${tokens.global.md.sys.typescale["section-title"].value.fontSize}px`,
-  //     lineHeight:
-  //       `${tokens.global.md.sys.typescale["section-title"].value.lineHeight}px` ===
-  //       "auto"
-  //         ? "auto"
-  //         : `${tokens.global.md.sys.typescale["section-title"].value.lineHeight}px`,
-  //     letterSpacing: `${tokens.global.md.sys.typescale["section-title"].value.letterSpacing}px`,
-  //     paragraphSpacing: `${tokens.global.md.sys.typescale["section-title"].value.paragraphSpacing}px`,
-  //   },
-  //   pageTitle: {
-  //     fontFamily: tokens.global.md.sys.typescale["page-title"].value.fontFamily,
-  //     fontWeight: 500,
-  //     fontSize: `${tokens.global.md.sys.typescale["page-title"].value.fontSize}px`,
-  //     lineHeight:
-  //       `${tokens.global.md.sys.typescale["page-title"].value.lineHeight}px` ===
-  //       "auto"
-  //         ? "auto"
-  //         : `${tokens.global.md.sys.typescale["page-title"].value.lineHeight}px`,
-  //     letterSpacing: `${tokens.global.md.sys.typescale["page-title"].value.letterSpacing}px`,
-  //     paragraphSpacing: `${tokens.global.md.sys.typescale["page-title"].value.paragraphSpacing}px`,
-  //   },
-  //   progressBarText: {
-  //     fontFamily:
-  //       tokens.global.md.sys.typescale["progress-bar-text"].value.fontFamily,
-  //     fontWeight: 600,
-  //     fontSize: `${tokens.global.md.sys.typescale["progress-bar-text"].value.fontSize}px`,
-  //     lineHeight:
-  //       `${tokens.global.md.sys.typescale["progress-bar-text"].value.lineHeight}px` ===
-  //       "auto"
-  //         ? "auto"
-  //         : `${tokens.global.md.sys.typescale["progress-bar-text"].value.lineHeight}px`,
-  //     letterSpacing: `${tokens.global.md.sys.typescale["progress-bar-text"].value.letterSpacing}px`,
-  //     paragraphSpacing: `${tokens.global.md.sys.typescale["progress-bar-text"].value.paragraphSpacing}px`,
-  //   },
-  //   userInstructions: {
-  //     fontFamily:
-  //       tokens.global.md.sys.typescale["user-instructions"].value.fontFamily,
-  //     fontWeight: 400,
-  //     fontSize: `${tokens.global.md.sys.typescale["user-instructions"].value.fontSize}px`,
-  //     lineHeight:
-  //       `${tokens.global.md.sys.typescale["user-instructions"].value.lineHeight}px` ===
-  //       "auto"
-  //         ? "auto"
-  //         : `${tokens.global.md.sys.typescale["user-instructions"].value.lineHeight}px`,
-  //     letterSpacing: `${tokens.global.md.sys.typescale["user-instructions"].value.letterSpacing}px`,
-  //     paragraphSpacing: `${tokens.global.md.sys.typescale["user-instructions"].value.paragraphSpacing}px`,
-  //   },
-  //   userInstructionTitles: {
-  //     fontFamily:
-  //       tokens.global.md.sys.typescale["user-instruction-titles"].value
-  //         .fontFamily,
-  //     fontWeight: 600,
-  //     fontSize: `${tokens.global.md.sys.typescale["user-instruction-titles"].value.fontSize}px`,
-  //     lineHeight:
-  //       `${tokens.global.md.sys.typescale["user-instruction-titles"].value.lineHeight}px` ===
-  //       "auto"
-  //         ? "auto"
-  //         : `${tokens.global.md.sys.typescale["user-instruction-titles"].value.lineHeight}px`,
-  //     letterSpacing: `${tokens.global.md.sys.typescale["user-instruction-titles"].value.letterSpacing}px`,
-  //     paragraphSpacing: `${tokens.global.md.sys.typescale["user-instruction-titles"].value.paragraphSpacing}px`,
-  //   },
-  //   buttonText: {
-  //     fontFamily:
-  //       tokens.global.md.sys.typescale["button-text"].value.fontFamily,
-  //     fontWeight: 500,
-  //     fontSize: `${tokens.global.md.sys.typescale["button-text"].value.fontSize}px`,
-  //     lineHeight:
-  //       `${tokens.global.md.sys.typescale["button-text"].value.lineHeight}px` ===
-  //       "auto"
-  //         ? "auto"
-  //         : `${tokens.global.md.sys.typescale["button-text"].value.lineHeight}px`,
-  //     letterSpacing: `${tokens.global.md.sys.typescale["button-text"].value.letterSpacing}px`,
-  //     paragraphSpacing: `${tokens.global.md.sys.typescale["button-text"].value.paragraphSpacing}px`,
-  //   },
-  //   tutorialTitle: {
-  //     fontFamily:
-  //       tokens.global.md.sys.typescale["tutorial-title"].value.fontFamily,
-  //     fontWeight: 500,
-  //     fontSize: `${tokens.global.md.sys.typescale["tutorial-title"].value.fontSize}px`,
-  //     lineHeight:
-  //       `${tokens.global.md.sys.typescale["tutorial-title"].value.lineHeight}px` ===
-  //       "auto"
-  //         ? "auto"
-  //         : `${tokens.global.md.sys.typescale["tutorial-title"].value.lineHeight}px`,
-  //     letterSpacing: `${tokens.global.md.sys.typescale["tutorial-title"].value.letterSpacing}px`,
-  //     paragraphSpacing: `${tokens.global.md.sys.typescale["tutorial-title"].value.paragraphSpacing}px`,
-  //   },
-  // },
+  typography: {
+    fontFamily: "Poppins, sans-serif",
+  },
   components: {
-    MuiCssBaseline: {
-      styleOverrides: {
-        "@global": {
-          "@font-face": [Poppins],
-        },
-      },
-    },
     MuiButton: {
       styleOverrides: {
         root: {
